Extract LearnPageComponent into a separate file
LearnPageComponent was causing Vercel build to fail due to it being
exported from a Next.js page.
